+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Added
 
 ### Changed
+- Enable SSL if the keystore password is set, remove the `ssl.enabled` property  
 
 ### Deprecated